モジュール化を更に進めたバージョン
ファイル操作系の関数群とアドレス操作系の関数群をそれぞれ別モジュールに分離
