+++ README.md
@@ -1,4 +1,5 @@
-# 住所録第5版
+# 住所録第6版
 
-住所録を管理するためのプログラムの第5版です。
-モジュール化を実施し、表示系の関数群を別モジュールとして切り出しました。
+住所録を管理するためのプログラムの第6版です。
+さらにモジュール化を実施し、ファイル操作系とアドレス操作系の
+関数群を別モジュールとして切り出しました。
