+++ denden.h
@@ -1,7 +1,7 @@
 /*
 **
 **	例題5：
-**		住所録管理表・第5版
+**		住所録管理表・第6版
 **
 **		file name：	denden.h
 **
